Patch received from Christian Kellermann (@ckeen).
Without this diff the warnings in weechat_otr.py:1203 and following are
misleading as the code path continues on and despite the promise not to
send any unencrypted text over the wire the message is sent anyway.

So this patch sets the result back to the empty string and does not
return with the current block by putting the following instruction into
an else branch.