weechat_otr.py: don't OTR with Anope services
This adds all services supported by Anope into the services list. They
don't handle receiving the tag correctly.

Atheme must be added separately as it has more services and there is no
clear list of those anywhere other than example config files.

This list of Anope services is based on http://anope.org/about.php and
StatServ was said to be supported at IRC and it was said to be too new
one to appear on that page.

ref: #114