Add otr.general.defaultkey option to share a key across accounts.
The previous (and still the default) behavior is to generate a new
private key per account. If you generate a key in this way and then
'/set otr.general.defaultkey nick@server' (nick@server is the account
of your existing key), new accounts will get this key as their private
key.

This is useful if you use key fingerprints to prove your identity
and want it to be the same on many different nicks and servers.

[closes #70]