README.md: add more installation methods for potr
Package manager would possibly be preferred to pip and if pip is
installed, installing the wheel module will make it a lot faster.

I also added Arch for potr requirements, because unlike most of other
distributions it uses Python 3 by default (which is why PEP 394 was
made).

No code changes so,
[CI SKIP]